requirements.txt
outlines
openai
assemblyai
python-dotenv
transformers
openpyxl